refactor: add indexes for imports
